feat: ✨ add concluding remarks (#105)
## Description

This PR adds some concluding remarks, mostly summarising what we’ve done
throughout the workshop + a what’s next section with links to other
`rostools` workshops. I’ve added a TODO on adding more big picture
content, bc I thought it would make sense to wait until the introductory
bigger picture has been made; then this conclusion can recap some of
that.

I’ve also moved the discussion activity on how they can use GitHub in
their own projects to this session. Since they will know about issues
and contributors as well at this point (as opposed to in the Working
with files session), I thought it might be nice to move it to the end of
the workshop. But if you disagree, we can revert that commit.

Closes #80 

<!-- Please delete as appropriate: -->
This PR needs an in-depth review.

## Checklist

- [X] Formatted Markdown
- [X] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>