feat: ✨ introduction slides (#173)
## Description

This is a quite quick and rough draft of what the introductory slides
could contain.

I know that I haven't set up the slides to use the `r3-theme`, but we
can do that afterwards (if you think we should keep these slides).

~~Also, the font awesome icons don't work in the slides 🤔~~ Edit: moved
schedule back to `schedule.qmd`, so this is not an issue anymore.

I'm really not set on this, so let me know what you think.

<!-- Please delete as appropriate: -->
This PR needs an in-depth review.

## Checklist

- [X] Ran `just run-all`

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Luke W. Johnston <lwjohnst@gmail.com>
Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>